# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Apixies.io API Documentation",
        "description": "API documentation for Apixies.io - A collection of utility APIs for developers",
        "contact": {
            "name": "Apixies.io Support",
            "email": "support@apixies.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://apixies.io",
            "description": "Production API Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Development API Server"
        }
    ],
    "paths": {
        "/api/v1/inspect-email": {
            "get": {
                "tags": [
                    "inspector"
                ],
                "summary": "Email Inspector",
                "description": "Inspect email address details including format validation, MX record checks, and disposable email detection",
                "operationId": "inspectEmail",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email address to inspect",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email",
                            "example": "user@example.com"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email inspection successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "X-API-KEY": []
                    }
                ]
            }
        },
...
        "/api/v1/html-to-pdf": {
            "post": {
                "tags": [
                    "converter"
                ],
                "summary": "HTML to PDF Converter",
                "description": "Convert HTML content to a PDF document with proper formatting and styling",
                "operationId": "htmlToPdf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "html"
                                ],
                                "properties": {
                                    "html": {
                                        "type": "string",
                                        "example": "<h1>Hello World</h1><p>Sample content</p>"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PDF document generated successfully",
                        "content": {
                            "application/pdf": {}
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "X-API-KEY": []
                    }
                ]
            }
        },
        "/api/v1/ip-geolocation": {
            "get": {
                "tags": [
                    "inspector"
                ],
                "summary": "IP Geolocation",
                "description": "Convert IP addresses to location data including country, city, coordinates, timezone, and ISP information",
                "operationId": "ipGeolocation",
                "parameters": [
                    {
                        "name": "ip",
                        "in": "query",
                        "description": "IP address to geolocate",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "8.8.8.8"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "IP geolocation retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid IP address",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "X-API-KEY": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "http_code": {
                        "type": "integer",
                        "example": 200
                    },
                    "code": {
                        "type": "string",
                        "example": "SUCCESS"
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation completed successfully"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "http_code": {
                        "type": "integer",
                        "example": 400
                    },
                    "code": {
                        "type": "string",
                        "example": "VALIDATION_ERROR"
                    },
                    "message": {
                        "type": "string",
                        "example": "The request data is invalid"
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "X-API-KEY": {
                "type": "apiKey",
                "description": "API Key for authentication",
                "name": "X-API-KEY",
                "in": "header"
            },
            "apiKey": {
                "type": "apiKey",
                "description": "API key for authentication",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "apiKey": []
        }
    ]
}